// frontend/mms-Admin/src/services/redux/types/notification.ts
export interface Notification {
    userId?: string;
    allNotificationsEmail: boolean;
    programsEmail: boolean;
    tasksEmail: boolean;
    approvalRequestsEmail: boolean;
    reportsEmail: boolean;
    commentsOnMyPostsEmail: boolean;
    postsEmail: boolean;
    commentsEmail: boolean;
    mentionsEmail: boolean;
    directMessagesEmail: boolean;
    allNotificationsInApp: boolean;
    programsInApp: boolean;
    tasksInApp: boolean;
    approvalRequestsInApp: boolean;
    reportsInApp: boolean;
    commentsOnMyPostsInApp: boolean;
    postsInApp: boolean;
    commentsInApp: boolean;
    mentionsInApp: boolean;
    directMessagesInApp: boolean;
  }

  